annonymous flow issue fixed

+++ app/helpers/deal_helper.rb
@@ -6,7 +6,7 @@ module DealHelper
   end
 
   def modify(deal)
-    if current_user.is_vendor && deal.eatery.user.id == current_user.id
+    if current_user. present? && current_user.is_vendor && deal.eatery.user.id == current_user.id
       link_to(image_tag("edit.png", title: "edit"), edit_vendor_deal_path(id: deal.id)) +
       link_to(image_tag("delete.png", title: "delete"), vendor_deal_path(id: deal.id), method: :delete)
     end
